+++ src/trx/game/output/uniforms.c
@@ -36,8 +36,10 @@
     X_DECLARE_MEMBER(int, reflections_enabled)                                 \
     X_DECLARE_MEMBER(int, textures_enabled)                                    \
     X_DECLARE_MEMBER(int, vertex_snap_enabled)                                 \
+    X_DECLARE_MEMBER(int, dither_mode)                                         \
     X_DECLARE_MEMBER(int, tr_version)                                          \
-    X_DECLARE_MEMBER(float, uv_scroll_tick)
+    X_DECLARE_MEMBER(float, uv_scroll_tick)                                    \
+    X_DECLARE_MEMBER(int, _pad, [3])
 
 #pragma pack(push, 4)
 typedef struct {
@@ -155,6 +157,7 @@ void Output_Uniforms_UploadGeneral(const OUTPUT_UNIFORMS *const uniforms)
         .trapezoid_filter_enabled = g_Config.rendering.enable_trapezoid_filter,
         .reflections_enabled = g_Config.visuals.enable_reflections,
         .vertex_snap_enabled = g_Config.rendering.enable_vertex_snap,
+        .dither_mode = g_Config.rendering.dither_mode,
         .fog_distance = {Output_GetFogStart(), Output_GetFogEnd()},
         .fog_color = {
             Output_GetFogColor().r,